requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
a2a-sdk[http-server]>=0.2.0
uvicorn>=0.30.0
